Move data_analysis to deltakit_core
