+++ deltakit-core/src/deltakit_core/data_formats/_data_analysis.py
